+++ libgearman/protocol/submit.cc
@@ -136,7 +136,7 @@ gearman_return_t submit_epoch(gearman_task_st& task,
   args_size[1]= task.unique_length +1;
 
   char time_string[30];
-  int length= snprintf(time_string, sizeof(time_string), "%lld", static_cast<long long>(when));
+  int length= snprintf(time_string, sizeof(time_string), "%" PRIu64, static_cast<int64_t>(when));
   args[2]= time_string;
   args_size[2]= length +1;
